Atualizar contrato

+++ Faucet_v2_aula06.sol
@@ -7,6 +7,7 @@
 
     /// @author Vitor Giacometti
     /// @title Estudando o conceito para Faucet
+    /// Contrato 0xE4B5adb8e3E7d107b49c2706E5dAA6BE9e687ca4
     contract Faucet_v2 {
 
         mapping(address=>uint8) public viewContrato;
@@ -18,7 +19,7 @@
         // @notice atualizar a quantidade de views em um contrato
         // @dev incrementar o contrato incrementa o valor do acumulador a um endereco ethereum.
         // @return valor atual do valorTotal
-        function atribuirValor() public returns (uint256) {
+        function atribuirValor() public returns (uint8) {
             valorTotal++;
             viewContrato[msg.sender] = valorTotal;
             emit viewChegouNoLimite(msg.sender, valorTotal);
